autogenerate FHIR API REGEX
